Added support for packaging DNX and NuGet specifications (or MSBuild projects).
Currently, only NuGet is being tested as I broke the chain into the test/src/condo.project folder from the root make.shade. We need to make the root make.shade instantiate a new instance of Sake using the nupkg that is created from the build -- which is the only viable way to execute the tests.

Related Work Items: #11