fix: resolve issue where build quality could be incorrect
When a CI server set the Build variable to a non-abbreviated form (such as refs/heads/develop), the build quality could be mis-identified as 'alpha'. We verify the start of a string as a match to a well-known branch, which required the abbreviated form previously. This has been resolved whereby the branch name is always abbreviated.