revert(semver): revert change to major prerelease
A previous change was introduced to condo to increment the minor segment of a version if a breaking change was detected,
which was intended to allow increments to continue to occur in feature branches after merging to the develop branch. For
instance:

Version: 1.0.0

A new feature is introduced with a breaking change:

Version: 2.0.0-alpha-01 (feature branch)

Feature is merged to develop:

Version: 2.0.0-beta-02 (develop branch)

A new feature is introduced in a separate feature branch:

Version: 2.0.0-alpha-03

In this scenario, a version range for NuGet packages of: [2.0.0-*, 3.0.0), would never pull 2.0.0-alpha-03 as
2.0.0-beta-02 is "newer" according to semver. Incrementing the minor version instead of the major version for all
prerelease builds was intended to make it easier to consume prerelease (alpha) packages with NuGet version ranges, but
it does break semver rules.

This commit reverts this in favor of strict semver compliance as described in the previous scenario.