fix(dotnet): save assembly info after prepare (#171)
* make `SaveAssemblyInfo` target fire after the prepare target

NOTE:
A previous change to condo moved when and how the dotnet project metadata was captured in order to improve performance in scenarios where a dotnet build would not execute, such as a clean target. This inadvertently stopped condo from writing the generated assembly info since the paths were not yet discovered during the version target.